requirement.txt
Django==1.11
djangorestframework==3.7.7
pkg-resources==0.0.0
psycopg2==2.7.3.2
pytz==2017.3
redis==2.10.6
